+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.1] - 2018-07-25
+### Updated
+- prefix function outside namespace for: `ContainerAwareTrait`, `BaseModel`, `Init` & `AbstractPlugin`. 
+- Change HttpFoundationRequestInterface::setRequest and `HttpFoundationRequestTrait` to return `self` instead of `parent`.
+
 ## [1.2.0] - 2018-06-07
 ### Added
 - HttpFoundation as composer suggestion.